requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.3.1
django-requestlogs==0.8
django-session-timeout==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
